refactor page, column readers and writers to use generics in order to reduce the amount of code being repeated
