+++ page_test.go
@@ -1,3 +1,5 @@
+//go:build !go1.18
+
 package parquet_test
 
 import (
